requirements.txt
ultralytics==8.2.0
opencv-python==4.9.0.80
fastapi==0.111.0
uvicorn==0.30.0
redis==5.0.4
psycopg2-binary==2.9.9
requests==2.32.2
python-dotenv==1.0.1
osmnx==1.9.3
networkx==3.3
twilio==9.1.0
roboflow==1.1.30
numpy==1.26.4
Pillow==10.3.0